requirements.txt
groq==0.12.0
psycopg2==2.9.10
pydantic==2.9.2
pydantic_core==2.23.4
pytest==8.3.3
python-dotenv==1.0.1
PyYAML==6.0.2
typing_extensions==4.12.2
